Signal to Noise (SN) eigenspectra produced but test results show values ranging from ~-.2 to ~.2.  Negative values indicate a problem here.
Removed a bug from make_Cmatrix.makeCmatrix: the forSMICA parameter that
was present there caused incorrect beamsmoothing to be applied.  This
was probably due to a sloppy translation from a routine in
filter_map.pro in IDL to makeCmatrix in python. Dividing by beam(5') has
been removed.  All C-matrices that used this feature are now obsolete.