1. Concluded testing of S/N eigenmode decomposition.  Rewrote parts of SN_mode_filter to demonstrate the technique.  Method looks very promising as only about 300 of 6000 modes will be needed in calculation.
2. New file mask_check.py runs through a series of masks based on
various sizes of apertures around the GNS coordinates.  Testing has
shown a trend when increasing aperture size to decrease the amplitude of
the template fit while only minimally changing the error estimate.  This
suggests that there may be negative amplitude compensations at higher
radii.  This also suggests that highpass filtering of the ISW template map using
the same method as the full sky simulations is not the correct method.

3. Consolidated testing filenames into template_fit.getFilenames.  Moved
data files to less confusing locations.  Expanded choice of ISW files to
include one based on PSG's figure 1, which has a different M(R) plot for
R=60 and R=100 Mpc/h than I calculate with ISWprofile.overmass.

4. Changed filter_map.pro (in IDL files) to have option for non-SMICA
smoothing.  Used this to re-do all smoothing on the 6 ISW files in
template_fit.getFilenames.